+++ ethos-frontend/src/contexts/AuthContextBase.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import type { AuthContextType } from '../types/authTypes';
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
