+++ ethos-frontend/src/contexts/ThemeHelpers.ts
@@ -1,0 +1,8 @@
+export type Theme = 'light' | 'dark' | 'system';
+
+export const getSystemTheme = (): 'light' | 'dark' =>
+  typeof window !== 'undefined' &&
+  window.matchMedia('(prefers-color-scheme: dark)').matches
+    ? 'dark'
+    : 'light';
+
